test: password prompt
