+++ jest.config.js
@@ -10,5 +10,5 @@ export default {
 		'^(\\.{1,2}/.*)\\.js$': '$1',
 	},
 	testRegex: ['__tests__/.+(spec|test).ts'],
-	setupFiles: ['<rootDir>/packages/core/__tests__/setup-tests.ts'],
+	setupFiles: ['<rootDir>/setup.tests.ts'],
 };
