+++ packages/prompts/src/index.ts
@@ -2,172 +2,15 @@ import {
 	ConfirmPrompt,
 	GroupMultiSelectPrompt,
 	MultiSelectPrompt,
-	PasswordPrompt,
 	SelectKeyPrompt,
-	SelectPrompt,
-	type State,
-	TextPrompt,
-	block,
-	isCancel,
+	SelectPrompt
 } from '@clack/core';
-import isUnicodeSupported from 'is-unicode-supported';
 import color from 'picocolors';
 import { cursor, erase } from 'sisteransi';
 
 export { isCancel, mockPrompt, setGlobalAliases } from '@clack/core';
-
-const unicode = isUnicodeSupported();
-const s = (c: string, fallback: string) => (unicode ? c : fallback);
-export const S_STEP_ACTIVE = s('◆', '*');
-export const S_STEP_CANCEL = s('■', 'x');
-export const S_STEP_ERROR = s('▲', 'x');
-export const S_STEP_SUBMIT = s('◇', 'o');
-
-export const S_BAR_START = s('┌', 'T');
-export const S_BAR = s('│', '|');
-export const S_BAR_END = s('└', '—');
-
-export const S_RADIO_ACTIVE = s('●', '>');
-export const S_RADIO_INACTIVE = s('○', ' ');
-export const S_CHECKBOX_ACTIVE = s('◻', '[•]');
-export const S_CHECKBOX_SELECTED = s('◼', '[+]');
-export const S_CHECKBOX_INACTIVE = s('◻', '[ ]');
-export const S_PASSWORD_MASK = s('▪', '•');
-
-export const S_BAR_H = s('─', '-');
-export const S_CORNER_TOP_RIGHT = s('╮', '+');
-export const S_CONNECT_LEFT = s('├', '+');
-export const S_CORNER_BOTTOM_RIGHT = s('╯', '+');
-
-export const S_INFO = s('●', '•');
-export const S_SUCCESS = s('◆', '*');
-export const S_WARN = s('▲', '!');
-export const S_ERROR = s('■', 'x');
-
-export const symbol = (state: State) => {
-	switch (state) {
-		case 'initial':
-		case 'active':
-			return color.cyan(S_STEP_ACTIVE);
-		case 'cancel':
-			return color.red(S_STEP_CANCEL);
-		case 'error':
-			return color.yellow(S_STEP_ERROR);
-		case 'submit':
-			return color.green(S_STEP_SUBMIT);
-	}
-};
-
-interface LimitOptionsParams<TOption> {
-	options: TOption[];
-	maxItems: number | undefined;
-	cursor: number;
-	style: (option: TOption, active: boolean) => string;
-}
-
-const limitOptions = <TOption>(params: LimitOptionsParams<TOption>): string[] => {
-	const { cursor, options, style } = params;
-
-	const paramMaxItems = params.maxItems ?? Number.POSITIVE_INFINITY;
-	const outputMaxItems = Math.max(process.stdout.rows - 4, 0);
-	// We clamp to minimum 5 because anything less doesn't make sense UX wise
-	const maxItems = Math.min(outputMaxItems, Math.max(paramMaxItems, 5));
-	let slidingWindowLocation = 0;
-
-	if (cursor >= slidingWindowLocation + maxItems - 3) {
-		slidingWindowLocation = Math.max(Math.min(cursor - maxItems + 3, options.length - maxItems), 0);
-	} else if (cursor < slidingWindowLocation + 2) {
-		slidingWindowLocation = Math.max(cursor - 2, 0);
-	}
-
-	const shouldRenderTopEllipsis = maxItems < options.length && slidingWindowLocation > 0;
-	const shouldRenderBottomEllipsis =
-		maxItems < options.length && slidingWindowLocation + maxItems < options.length;
-
-	return options
-		.slice(slidingWindowLocation, slidingWindowLocation + maxItems)
-		.map((option, i, arr) => {
-			const isTopLimit = i === 0 && shouldRenderTopEllipsis;
-			const isBottomLimit = i === arr.length - 1 && shouldRenderBottomEllipsis;
-			return isTopLimit || isBottomLimit
-				? color.dim('...')
-				: style(option, i + slidingWindowLocation === cursor);
-		});
-};
-
-export function formatPlaceholder(placeholder: string | undefined): string {
-	return placeholder
-		? color.inverse(placeholder[0]) + color.dim(placeholder.slice(1))
-		: color.inverse(color.hidden('_'));
-}
-
-export interface TextOptions {
-	message: string;
-	placeholder?: string;
-	defaultValue?: string;
-	initialValue?: string;
-	validate?: (value: string) => string | undefined;
-}
-export const text = (opts: TextOptions) => {
-	return new TextPrompt({
-		validate: opts.validate,
-		placeholder: opts.placeholder,
-		defaultValue: opts.defaultValue,
-		initialValue: opts.initialValue,
-		render() {
-			const title = `${color.gray(S_BAR)}\n${symbol(this.state)}  ${opts.message}\n`;
-			const placeholder = formatPlaceholder(opts.placeholder);
-			const value = !this.value ? placeholder : this.valueWithCursor;
-
-			switch (this.state) {
-				case 'error':
-					return `${title.trim()}\n${color.yellow(S_BAR)}  ${value}\n${color.yellow(
-						S_BAR_END
-					)}  ${color.yellow(this.error)}\n`;
-				case 'submit':
-					return `${title}${color.gray(S_BAR)}  ${color.dim(this.value || opts.placeholder)}`;
-				case 'cancel':
-					return `${title}${color.gray(S_BAR)}  ${color.strikethrough(
-						color.dim(this.value ?? '')
-					)}${this.value?.trim() ? `\n${color.gray(S_BAR)}` : ''}`;
-				default:
-					return `${title}${color.cyan(S_BAR)}  ${value}\n${color.cyan(S_BAR_END)}\n`;
-			}
-		},
-	}).prompt() as Promise<string | symbol>;
-};
-
-export interface PasswordOptions {
-	message: string;
-	mask?: string;
-	validate?: (value: string) => string | undefined;
-}
-export const password = (opts: PasswordOptions) => {
-	return new PasswordPrompt({
-		validate: opts.validate,
-		mask: opts.mask ?? S_PASSWORD_MASK,
-		render() {
-			const title = `${color.gray(S_BAR)}\n${symbol(this.state)}  ${opts.message}\n`;
-			const value = this.valueWithCursor;
-			const masked = this.maskedValue;
-
-			switch (this.state) {
-				case 'error':
-					return `${title.trim()}\n${color.yellow(S_BAR)}  ${masked}\n${color.yellow(
-						S_BAR_END
-					)}  ${color.yellow(this.error)}\n`;
-				case 'submit':
-					return `${title}${color.gray(S_BAR)}  ${color.dim(masked)}`;
-				case 'cancel':
-					return `${title}${color.gray(S_BAR)}  ${color.strikethrough(color.dim(masked ?? ''))}${
-						masked ? `\n${color.gray(S_BAR)}` : ''
-					}`;
-				default:
-					return `${title}${color.cyan(S_BAR)}  ${value}\n${color.cyan(S_BAR_END)}\n`;
-			}
-		},
-	}).prompt() as Promise<string | symbol>;
-};
+export { default as password, PasswordOptions } from './prompts/password';
+export { default as text, TextOptions } from './prompts/text';
 
 export interface ConfirmOptions {
 	message: string;
