+++ packages/prompts/src/prompts/text.ts
@@ -1,0 +1,42 @@
+import { TextPrompt } from '@clack/core';
+import color from 'picocolors';
+import { formatPlaceholder, symbol, S_BAR, S_BAR_END } from '../utils';
+
+export interface TextOptions {
+	message: string;
+	placeholder?: string;
+	defaultValue?: string;
+	initialValue?: string;
+	validate?: (value: string) => string | void;
+}
+
+const text = (opts: TextOptions) => {
+	return new TextPrompt({
+		validate: opts.validate,
+		placeholder: opts.placeholder,
+		defaultValue: opts.defaultValue,
+		initialValue: opts.initialValue,
+		render() {
+			const title = `${color.gray(S_BAR)}\n${symbol(this.state)}  ${opts.message}\n`;
+			const placeholder = formatPlaceholder(opts.placeholder);
+			const value = !this.value ? placeholder : this.valueWithCursor;
+
+			switch (this.state) {
+				case 'error':
+					return `${title.trim()}\n${color.yellow(S_BAR)}  ${value}\n${color.yellow(
+						S_BAR_END
+					)}  ${color.yellow(this.error)}\n`;
+				case 'submit':
+					return `${title}${color.gray(S_BAR)}  ${color.dim(this.value || opts.placeholder)}`;
+				case 'cancel':
+					return `${title}${color.gray(S_BAR)}  ${
+						this.value ? color.strikethrough(color.dim(this.value)) : ''
+					}${this.value ? '\n' + color.gray(S_BAR) : ''}`;
+				default:
+					return `${title}${color.cyan(S_BAR)}  ${value}\n${color.cyan(S_BAR_END)}\n`;
+			}
+		},
+	}).prompt() as Promise<string | symbol>;
+};
+
+export default text;
