+++ packages/prompts/src/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './format';
+export * from './symbol';
